src/security/auth.py
import jwt

SECRET_KEY = "supersecret"
ALGORITHM = "HS256"


def create_jwt_token(data: dict):
    return jwt.encode(data, SECRET_KEY, algorithm=ALGORITHM)


def decode_jwt_token(token: str):
    return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
